porting more things from commit 0d1a038

+++ src/Mewdeko/Services/GreetSettingsService.cs
@@ -1,4 +1,4 @@
-using Mewdeko.Common.ModuleBehaviors;
+﻿using Mewdeko.Common.ModuleBehaviors;
 using Mewdeko.Services.Common;
 using Mewdeko.Services.Settings;
 using Serilog;
@@ -89,8 +89,7 @@ public class GreetSettingsService : INService, IReadyExecutor
             {
                 var toDelete = await chan.SendMessageAsync(plainText, embeds: embed, components: components?.Build())
                     .ConfigureAwait(false);
-                if (conf.BoostMessageDeleteAfter > 0)
-                    toDelete.DeleteAfter(conf.BoostMessageDeleteAfter);
+                if (conf.BoostMessageDeleteAfter > 0) toDelete.DeleteAfter(conf.BoostMessageDeleteAfter);
             }
             catch (Exception ex)
             {
@@ -104,8 +103,7 @@ public class GreetSettingsService : INService, IReadyExecutor
             {
                 var toDelete = await chan.SendMessageAsync(msg).ConfigureAwait(false);
 
-                if (conf.BoostMessageDeleteAfter > 0)
-                    toDelete.DeleteAfter(conf.BoostMessageDeleteAfter);
+                if (conf.BoostMessageDeleteAfter > 0) toDelete.DeleteAfter(conf.BoostMessageDeleteAfter);
             }
             catch (Exception ex)
             {
@@ -160,12 +158,11 @@ public class GreetSettingsService : INService, IReadyExecutor
             var user = usr as SocketGuildUser;
             var conf = await GetOrAddSettingsForGuild(guild.Id);
 
-            if (!conf.SendChannelByeMessage)
-                return;
+            if (!conf.SendChannelByeMessage) return;
 
             if ((await guild.GetTextChannelsAsync()).SingleOrDefault(c =>
                     c.Id == conf.ByeMessageChannelId) is not
-                    { } channel) //maybe warn the server owner that the channel is missing
+                { } channel) //maybe warn the server owner that the channel is missing
             {
                 return;
             }
@@ -194,9 +191,9 @@ public class GreetSettingsService : INService, IReadyExecutor
 
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         await uow.SaveChangesAsync().ConfigureAwait(false);
-        return conf.SendBoostMessage;
+        return false.ParseBoth(conf.SendBoostMessage.ToString());
     }
 
     public async Task SetBoostDel(ulong guildId, int timer)
@@ -207,7 +204,7 @@ public class GreetSettingsService : INService, IReadyExecutor
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
         conf.BoostMessageDeleteAfter = timer;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
@@ -221,15 +218,17 @@ public class GreetSettingsService : INService, IReadyExecutor
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        var enabled = conf.SendBoostMessage = value ?? !conf.SendBoostMessage;
+        var enabledLong = conf.SendBoostMessage == 0 ? 1 : 0;
+        var isEnabled = (value.HasValue) ? (value.Value ? 1L : 0L) : enabledLong;
+        conf.SendBoostMessage = isEnabled;
         conf.BoostMessageChannelId = channelId;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return enabled;
+        return isEnabled != 0;
     }
 
     public async Task SetWebGreetUrl(IGuild guild, string url)
@@ -238,7 +237,7 @@ public class GreetSettingsService : INService, IReadyExecutor
         var gc = await uow.ForGuildId(guild.Id, set => set);
         gc.GreetHook = url;
         await uow.SaveChangesAsync().ConfigureAwait(false);
-        gss.UpdateGuildConfig(guild.Id, gc);
+        await gss.UpdateGuildConfig(guild.Id, gc);
     }
 
     public async Task SetWebLeaveUrl(IGuild guild, string url)
@@ -247,7 +246,7 @@ public class GreetSettingsService : INService, IReadyExecutor
         var gc = await uow.ForGuildId(guild.Id, set => set);
         gc.LeaveHook = url;
         await uow.SaveChangesAsync().ConfigureAwait(false);
-        gss.UpdateGuildConfig(guild.Id, gc);
+        await gss.UpdateGuildConfig(guild.Id, gc);
     }
 
     public async Task<string> GetDmGreetMsg(ulong id)
@@ -296,8 +295,7 @@ public class GreetSettingsService : INService, IReadyExecutor
                     var toDelete = await channel
                         .SendMessageAsync(plainText, embeds: embed, components: components?.Build())
                         .ConfigureAwait(false);
-                    if (conf.AutoDeleteByeMessagesTimer > 0)
-                        toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
+                    if (conf.AutoDeleteByeMessagesTimer > 0) toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
                 }
                 else
                 {
@@ -327,8 +325,7 @@ public class GreetSettingsService : INService, IReadyExecutor
                 if (string.IsNullOrEmpty(lh) || lh == 0.ToString())
                 {
                     var toDelete = await channel.SendMessageAsync(msg.SanitizeMentions()).ConfigureAwait(false);
-                    if (conf.AutoDeleteByeMessagesTimer > 0)
-                        toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
+                    if (conf.AutoDeleteByeMessagesTimer > 0) toDelete.DeleteAfter(conf.AutoDeleteByeMessagesTimer);
                 }
                 else
                 {
@@ -350,9 +347,9 @@ public class GreetSettingsService : INService, IReadyExecutor
 
     private Task GreetUsers(GreetSettings conf, ITextChannel channel, IGuildUser user) => GreetUsers(conf, channel,
         new[]
-    {
-        user
-    });
+        {
+            user
+        });
 
     private async Task GreetUsers(GreetSettings conf, ITextChannel channel, IEnumerable<IGuildUser> users)
     {
@@ -456,8 +453,7 @@ public class GreetSettingsService : INService, IReadyExecutor
         else
         {
             var msg = rep.Replace(conf.DmGreetMessageText);
-            if (string.IsNullOrWhiteSpace(msg))
-                return true;
+            if (string.IsNullOrWhiteSpace(msg)) return true;
             try
             {
                 await channel.SendConfirmAsync(msg).ConfigureAwait(false);
@@ -520,8 +516,7 @@ public class GreetSettingsService : INService, IReadyExecutor
                 {
                     var channel = await user.CreateDMChannelAsync().ConfigureAwait(false);
 
-                    if (channel != null)
-                        await GreetDmUser(conf, user).ConfigureAwait(false);
+                    if (channel != null) await GreetDmUser(conf, user).ConfigureAwait(false);
                 }
             }
             catch
@@ -569,16 +564,16 @@ public class GreetSettingsService : INService, IReadyExecutor
         conf.ChannelByeMessageText = settings.ChannelByeMessageText?.SanitizeMentions();
 
         conf.AutoDeleteGreetMessagesTimer = settings.AutoDeleteGreetMessagesTimer;
-        conf.AutoDeleteGreetMessages = settings.AutoDeleteGreetMessagesTimer > 0;
+        conf.AutoDeleteGreetMessages = settings.AutoDeleteGreetMessagesTimer > 0 ? 1 : 0;
 
         conf.AutoDeleteByeMessagesTimer = settings.AutoDeleteByeMessagesTimer;
-        conf.AutoDeleteByeMessages = settings.AutoDeleteByeMessagesTimer > 0;
+        conf.AutoDeleteByeMessages = settings.AutoDeleteByeMessagesTimer > 0 ? 1 : 0;
 
         conf.GreetMessageChannelId = settings.GreetMessageChannelId;
         conf.ByeMessageChannelId = settings.ByeMessageChannelId;
 
-        conf.SendChannelGreetMessage = settings.SendChannelGreetMessage;
-        conf.SendChannelByeMessage = settings.SendChannelByeMessage;
+        conf.SendChannelGreetMessage = settings.SendChannelGreetMessage ? 1 : 0;
+        conf.SendChannelByeMessage = settings.SendChannelByeMessage ? 1 : 0;
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
@@ -592,7 +587,9 @@ public class GreetSettingsService : INService, IReadyExecutor
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        var enabled = conf.SendChannelGreetMessage = value ?? !conf.SendChannelGreetMessage;
+        var enabledLong = conf.SendChannelGreetMessage == 0 ? 1 : 0;
+        var isEnabled = (value.HasValue) ? (value.Value ? 1L : 0L) : enabledLong;
+        conf.SendChannelGreetMessage = isEnabled;
         conf.GreetMessageChannelId = channelId;
 
         var toAdd = GreetSettings.Create(conf);
@@ -600,7 +597,7 @@ public class GreetSettingsService : INService, IReadyExecutor
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return enabled;
+        return isEnabled != 0;
     }
 
 
@@ -615,27 +612,29 @@ public class GreetSettingsService : INService, IReadyExecutor
         var conf = await uow.ForGuildId(guildId, set => set);
         conf.ChannelGreetMessageText = message;
         var greetMsgEnabled = conf.SendChannelGreetMessage;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return greetMsgEnabled;
+        return false.ParseBoth(greetMsgEnabled.ToString());
     }
 
     public async Task<bool> SetGreetDm(ulong guildId, bool? value = null)
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        var enabled = conf.SendDmGreetMessage = value ?? !conf.SendDmGreetMessage;
-        gss.UpdateGuildConfig(guildId, conf);
+        var enabledLong = conf.SendDmGreetMessage == 0 ? 1 : 0;
+        var isEnabled = (value.HasValue) ? (value.Value ? 1L : 0L) : enabledLong;
+        conf.SendDmGreetMessage = isEnabled;
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return enabled;
+        return isEnabled != 0;
     }
 
 
@@ -650,28 +649,30 @@ public class GreetSettingsService : INService, IReadyExecutor
         var conf = await uow.ForGuildId(guildId, set => set);
         conf.DmGreetMessageText = message;
         var greetMsgEnabled = conf.SendDmGreetMessage;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return greetMsgEnabled;
+        return false.ParseBoth(greetMsgEnabled.ToString());
     }
 
     public async Task<bool> SetBye(ulong guildId, ulong channelId, bool? value = null)
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        var enabled = conf.SendChannelByeMessage = value ?? !conf.SendChannelByeMessage;
+        var enabledLong = conf.SendChannelByeMessage == 0 ? 1 : 0;
+        var isEnabled = (value.HasValue) ? (value.Value ? 1L : 0L) : enabledLong;
+        conf.SendChannelByeMessage = isEnabled;
         conf.ByeMessageChannelId = channelId;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return enabled;
+        return isEnabled != 0;
     }
 
 
@@ -686,13 +687,13 @@ public class GreetSettingsService : INService, IReadyExecutor
         var conf = await uow.ForGuildId(guildId, set => set);
         conf.ChannelByeMessageText = message;
         var byeMsgEnabled = conf.SendChannelByeMessage;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
         await uow.SaveChangesAsync().ConfigureAwait(false);
 
-        return byeMsgEnabled;
+        return false.ParseBoth(byeMsgEnabled.ToString());
     }
 
     public async Task SetByeDel(ulong guildId, int timer)
@@ -703,7 +704,7 @@ public class GreetSettingsService : INService, IReadyExecutor
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
         conf.AutoDeleteByeMessagesTimer = timer;
-        gss.UpdateGuildConfig(guildId, conf);
+        await gss.UpdateGuildConfig(guildId, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(guildId, toAdd, (_, _) => toAdd);
 
@@ -718,7 +719,7 @@ public class GreetSettingsService : INService, IReadyExecutor
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(id, set => set);
         conf.AutoDeleteGreetMessagesTimer = timer;
-        gss.UpdateGuildConfig(id, conf);
+        await gss.UpdateGuildConfig(id, conf);
         var toAdd = GreetSettings.Create(conf);
         GuildConfigsCache.AddOrUpdate(id, toAdd, (_, _) => toAdd);
 
@@ -731,28 +732,28 @@ public class GreetSettingsService : INService, IReadyExecutor
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        return conf.SendDmGreetMessage;
+        return false.ParseBoth(conf.SendDmGreetMessage.ToString());
     }
 
     public async Task<bool> GetGreetEnabled(ulong guildId)
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        return conf.SendChannelGreetMessage;
+        return false.ParseBoth(conf.SendChannelGreetMessage.ToString());
     }
 
     public async Task<bool> GetBoostEnabled(ulong guildId)
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        return conf.SendBoostMessage;
+        return false.ParseBoth(conf.SendBoostMessage.ToString());
     }
 
     public async Task<bool> GetByeEnabled(ulong guildId)
     {
         await using var uow = db.GetDbContext();
         var conf = await uow.ForGuildId(guildId, set => set);
-        return conf.SendChannelByeMessage;
+        return false.ParseBoth(conf.SendChannelByeMessage.ToString());
     }
 
     #endregion
@@ -816,15 +817,15 @@ public class GreetSettings
             AutoDeleteGreetMessagesTimer = g.AutoDeleteGreetMessagesTimer,
             GreetMessageChannelId = g.GreetMessageChannelId,
             ByeMessageChannelId = g.ByeMessageChannelId,
-            SendDmGreetMessage = g.SendDmGreetMessage,
+            SendDmGreetMessage = false.ParseBoth(g.SendDmGreetMessage.ToString()),
             DmGreetMessageText = g.DmGreetMessageText,
-            SendChannelGreetMessage = g.SendChannelGreetMessage,
+            SendChannelGreetMessage = false.ParseBoth(g.SendChannelGreetMessage.ToString()),
             ChannelGreetMessageText = g.ChannelGreetMessageText,
-            SendChannelByeMessage = g.SendChannelByeMessage,
+            SendChannelByeMessage = false.ParseBoth(g.SendChannelByeMessage.ToString()),
             ChannelByeMessageText = g.ChannelByeMessageText,
             BoostMessage = g.BoostMessage,
             BoostMessageChannelId = g.BoostMessageChannelId,
             BoostMessageDeleteAfter = g.BoostMessageDeleteAfter,
-            SendBoostMessage = g.SendBoostMessage
+            SendBoostMessage = false.ParseBoth(g.SendBoostMessage.ToString())
         };
 }
